# ggml : test dot product q4_0 x f32