@slaren 2023-04-18
I tested this a while ago because I was also very surprised that quantizing first is faster, but it is indeed faster. I even tried an AVX2 implementation of `ggml_vec_dot_q4_0_f32`. Here it is in case anyone is interested:

<details>

```C
static void ggml_vec_dot_q4_0_f32(const int n, float * restrict s, const void * restrict vx, const float * restrict y) {
    const int nb = n / QK;

    assert(n % QK == 0);
    assert(nb % 2 == 0);

    const block_q4_0 * restrict x = (const block_q4_0*)vx;

    __m256 acc = _mm256_setzero_ps();

    // Main loop
    for (int i = 0; i < nb; ++i) {
        const __m256 d_v = _mm256_broadcast_ss(&x[i].d);

        // Load 32x4-bit integers into 32x8-bit integers
        __m256i vx8 = bytesFromNibbles(x[i].qs);

        // Subtract 8 from the integers
        vx8 = _mm256_sub_epi8(vx8, _mm256_set1_epi8(8));

        // Convert to 16-bit int
        const __m256i vx16_lo = _mm256_cvtepi8_epi16(_mm256_extracti128_si256(vx8, 0));
        const __m256i vx16_hi = _mm256_cvtepi8_epi16(_mm256_extracti128_si256(vx8, 1));

        // Convert to 32-bit int -> float 32
        const __m256 vf[4] = {
            _mm256_cvtepi32_ps(_mm256_cvtepi16_epi32(_mm256_extracti128_si256(vx16_lo, 0))),
            _mm256_cvtepi32_ps(_mm256_cvtepi16_epi32(_mm256_extracti128_si256(vx16_lo, 1))),
            _mm256_cvtepi32_ps(_mm256_cvtepi16_epi32(_mm256_extracti128_si256(vx16_hi, 0))),
            _mm256_cvtepi32_ps(_mm256_cvtepi16_epi32(_mm256_extracti128_si256(vx16_hi, 1)))
        };

        // Scale and fma
        const float * yj = y + i*32;
        for (int j = 0; j < 4; j++) {
            const __m256 jx = _mm256_mul_ps(vf[j], d_v);
            const __m256 jy = _mm256_loadu_ps(yj + j*8);
            acc = _mm256_fmadd_ps(jx, jy, acc);
        }
    }

    // Return horizontal sum of the acc vector
    __m128 res = _mm256_extractf128_ps( acc, 1 );
    res = _mm_add_ps( res, _mm256_castps256_ps128( acc ) );
    res = _mm_add_ps( res, _mm_movehl_ps( res, res ) );
    res = _mm_add_ss( res, _mm_movehdup_ps( res ) );

    *s = _mm_cvtss_f32( res );
}
```
</details>