@ikawrakow 2023-04-20
The thing that I did not realize is that the number of values being quantized is many times less than the number of values used in the dot products (512X or more) . Which means that my simple test code from #1041 does not adequately measure the actual situation where the amount of time spent on quantization is small compared to the amount of time spent on dot products between quantized values. If I only consider the dot products, then indeed the quantized version is faster.

Sorry for wasting @ggerganov's time.